Feature/track crud

+++ src/main/java/io/ipme/neptunes/Controller/TrackController.java
@@ -3,11 +3,14 @@ package io.ipme.neptunes.Controller;
 import io.ipme.neptunes.Model.Track;
 import io.ipme.neptunes.Service.TrackService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
+@Valid
 @RestController
 public class TrackController {
 
@@ -15,26 +18,62 @@ public class TrackController {
     private TrackService trackService;
 
     @GetMapping("/tracks")
-    public List<Track> getAll(){
-      return trackService.findAll();
+    public ResponseEntity<List<Track>> getAll(){
+        try {
+            return ResponseEntity.ok().body(trackService.findAll());
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
     }
 
-    @GetMapping("/track/{id}")
-    public Optional<Track> getOne(@PathVariable Integer id){
-        return trackService.findOne(id);
+    @GetMapping("/tracks/{id}")
+    public ResponseEntity<Optional<Track>> getOne(@PathVariable Integer id){
+        try {
+            if (id != null) {
+                return ResponseEntity.ok().body(trackService.findOne(id));
+            }
+            return ResponseEntity.badRequest().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
     }
 
-    @PostMapping("/track/add")
-    public void createTrack(@RequestBody Track track) {
-        trackService.save(track);
-    }
-    @PostMapping("/track/remove/{id}")
-    public void createTrack(@RequestBody @PathVariable Integer id) {
-        trackService.remove(id);
+    @PostMapping("/tracks")
+    public ResponseEntity<String> createTrack(@RequestBody Track track) {
+        try {
+            if (track != null){
+                trackService.save(track);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
     }
 
-    @PutMapping("/track/update/{id}")
-    public void updateTrack(@RequestBody Track track, @PathVariable Integer id) {
-        trackService.update(track, id);
+    @DeleteMapping("/tracks/{id}")
+    public ResponseEntity<Track> removeTrack(@RequestBody @PathVariable Integer id) {
+        try {
+            if (id != null){
+                trackService.remove(id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.badRequest().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    @PatchMapping("/tracks/{id}")
+    public ResponseEntity<String> updateTrack(@RequestBody Track track, @PathVariable Integer id) {
+        try {
+            if (id != null && track != null){
+                trackService.update(track, id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.badRequest().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
     }
 }
